Use babel-core/register instead of babel-node CLI
